% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daylight_CBM.R
\name{daylight_CBM}
\alias{daylight_CBM}
\title{CBM Daylight Model}
\usage{
daylight_CBM(JDate, Latitude, p = 0.833)
}
\arguments{
\item{JDate}{Julian Day of the Year}

\item{Latitude}{Latitude in Degrees}

\item{p}{Daylength coefficient (Default: Constant based on Astronomical calendar)}
}
\value{
Daylight hours
}
\description{
Calculating the N daynight hours.
}
\details{
For details see Forsythe (1995) and Schoolfield (1982)
}
