% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daylight_N.R
\name{daylight_N}
\alias{daylight_N}
\title{N Daylight Hours Model}
\usage{
daylight_N(Date, Latitude)
}
\arguments{
\item{Date}{Date vector}

\item{Latitude}{Latitude in Degrees}
}
\value{
Daylight hours
}
\description{
Calculating the N daynight hours.
}
\details{
For details see Allen (1998)
}
