% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pet_hamon.R
\name{pet_hamon}
\alias{pet_hamon}
\title{Hamon Method for Potential Evapotranspiration}
\usage{
pet_hamon(par, JDate, Tavg, Latitude, dayFUN = sacramentoR::daylight_CBM)
}
\arguments{
\item{JDate}{Julian Date (Time-series or numeric)}

\item{Tavg}{Average Temperature (deg C) (Time-series or numeric)}

\item{Latitude}{Latitude (Degrees)}

\item{dayFUN}{Daylight hours function (Defaults to CBM Model)}

\item{Coeff}{Hamon PET Model Coefficient (Default: 1.2 from Lu (2005))}
}
\value{
Hamon Potential Evapotranspiration Value
}
\description{
Calculation of potential evapotranspiration based on the Hamon method
}
\details{
For details see Lu, Sun, McNulty, and Amatya (2005) and Haith and Shoemaker (1987)
}
