% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prmsRtosacsmaR.R
\name{prsmRtosacsmaR}
\alias{prsmRtosacsmaR}
\title{Translating PRMS Configuration Files to SAC-SMA-DS Format}
\usage{
prsmRtosacsmaR(prms.param,
  projection.from = "+proj=aea +lat_1=29.5 +lat_2=45.5 +lat_0=23 +lon_0=-96 +x_0=0 +y_0=0 +ellps=GRS80 +towgs84=1,1,-1,0,0,0,0 +units=m +no_defs",
  projection.to = "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs",
  elev_convert = TRUE, area_convert = TRUE)
}
\arguments{
\item{prms.param}{prmsR parameter variable (assumes units are in English, except for Elevation)}

\item{projection.from}{PROJ4 string for CRS (Default: CONUS Albers Projection)}

\item{projection.to}{PROJ4 string for CRS (Default: WGS84 Google Maps)}

\item{elev_convert}{Converting elevation from feet to meters}

\item{area_convert}{Converting area from acres to square meters}
}
\value{
output.configuration Outputs translated configurations
}
\description{
These translates prmsR's data format and PRMS's configuration files to SAC-SMA-DS use.
}
\details{
Note: This function is used specifically for California and could
}
