% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing_lohmann.R
\name{routing_lohmann}
\alias{routing_lohmann}
\title{Lohmann Routing Module}
\usage{
routing_lohmann(Param, directInflow, baseInflow, flowLength,
  Outlet = FALSE, WatershedCharacteristics = c(12, 96, 3600),
  preserveInput = FALSE, verbose = FALSE)
}
\arguments{
\item{Param}{routing module parameter list}

\item{directInflow}{Timeseries of Direct Inflow (mm) (xts object)}

\item{baseInflow}{Timeseries of Base Inflow (mm) (xts object)}

\item{flowLength}{Travel distance from catchment outlet to watershed outlet (meters)}

\item{Outlet}{Flag to determine if this is the outlet}

\item{WatershedCharacteristics}{Watershed Characteristics Vector c(KE, UH_DAY, DT) (not recommended to change this unless you know what you're doing)}

\item{preserveInput}{Flag to preserve the input as part of the output}

\item{verbose}{Verbose Flag}
}
\value{
A list of various time-series as an xts object
}
\description{
Calculates the Lohmann Routing model coupled with land surface parameterization
}
\details{
For details see Lohmann, Nolte-Holube, and Raschke (1996)
}
