% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_A_v.R
\name{utility_A_v}
\alias{utility_A_v}
\title{A_v Calculator}
\usage{
utility_A_v(Date, Latitude = 60)
}
\arguments{
\item{Date}{Date (as Date)}

\item{Latitude}{Latitude (Degrees) (Default value triggers calculation)}
}
\value{
A_v
}
\description{
Calculates the seasonal variation adjustment (A_v) based on Anderson 2006 equation 7
}
\details{
For details see Anderson (2006)
}
