% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_areaRatio.R
\name{utility_areaRatio}
\alias{utility_areaRatio}
\title{Area absolute to ratio of total area}
\usage{
utility_areaRatio(Area)
}
\arguments{
\item{Area}{A vector list of absolute area values}
}
\value{
A vector of ratio that all adds up to 1
}
\description{
Converts area from absolute values to a ratio of the total area
}
