% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_isLeapYear.R
\name{utility_isLeapYear}
\alias{utility_isLeapYear}
\title{Leap Year Checker}
\usage{
utility_isLeapYear(Year)
}
\arguments{
\item{Year}{Year}
}
\value{
True if leap year
}
\description{
Checks to see if it's a leap year
}
